+++ examples/data-sources/catalystcenter_sda_fabric_authentication_profile/data-source.tf
@@ -1,0 +1,3 @@
+data "catalystcenter_sda_fabric_authentication_profile" "example" {
+  id = "Global/Site1"
+}