+++ examples/resources/catalystcenter_sda_fabric_authentication_profile/resource.tf
@@ -1,0 +1,4 @@
+resource "catalystcenter_sda_fabric_authentication_profile" "example" {
+  site_name_hierarchy          = "Global/Site1"
+  authentication_template_name = "No Authentication"
+}