+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `catalystcenter_sda_fabric_site` resource and data source
 - Add `catalystcenter_template_version` resource and data source
+- Add `catalystcenter_sda_fabric_authentication_profile` resource and data source
 
 ## 0.1.4